Undo change: relocate Account to model_old. Create gen pkg, relocate new schema DTOs there.
