+++ common/src/main/java/com/microsoft/identity/common/internal/providers/microsoft/activedirectoryfederationservices/ActiveDirectoryFederationServices2012R2OAuth2Strategy.java
@@ -2,7 +2,7 @@ package com.microsoft.identity.common.internal.providers.microsoft.activedirecto
 
 import android.net.Uri;
 
-import com.microsoft.identity.common.model_old.Account;
+import com.microsoft.identity.common.Account;
 import com.microsoft.identity.common.internal.providers.oauth2.AccessToken;
 import com.microsoft.identity.common.internal.providers.oauth2.AuthorizationRequest;
 import com.microsoft.identity.common.internal.providers.oauth2.AuthorizationResponse;
